requirements.txt
pytesseract
opencv-python
Pillow
googletrans==4.0.0-rc1
